piprequirements.txt
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.2
Django==5.0.3
django-cors-headers==4.3.1
django-filter==23.5
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
msgpack==1.0.8
packaging==24.0
phonenumbers==8.13.31
pillow==10.2.0
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
pypng==0.20220715.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.1
requests==2.31.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
zope.interface==6.3
